+++ packages/actor-scraper/jsdom-scraper/src/internals/consts.ts
@@ -9,6 +9,7 @@ export interface Input {
     globs: GlobInput[];
     regexps: RegExpInput[];
     pseudoUrls: PseudoUrlInput[];
+    excludes: GlobInput[];
     keepUrlFragments: boolean;
     runScripts: boolean;
     showInternalConsole: boolean;