+++ packages/actor-scraper/playwright-scraper/src/internals/consts.ts
@@ -10,6 +10,7 @@ export interface Input {
     globs: GlobInput[];
     regexps: RegExpInput[];
     pseudoUrls: PseudoUrlInput[];
+    excludes: GlobInput[];
     linkSelector?: string;
     keepUrlFragments: boolean;
     pageFunction: string;
@@ -21,8 +22,10 @@ export interface Input {
     initialCookies: Parameters<Session['setCookies']>[0];
     launcher: 'chromium' | 'firefox' | 'webkit';
     useChrome: boolean;
+    maxScrollHeightPixels: number;
     ignoreSslErrors: boolean;
     ignoreCorsAndCsp: boolean;
+    closeCookieModals: boolean;
     downloadMedia: boolean;
     downloadCss: boolean;
     maxRequestRetries: number;