+++ packages/actor-scraper/puppeteer-scraper/src/internals/consts.ts
@@ -11,6 +11,7 @@ export interface Input {
     globs: GlobInput[];
     regexps: RegExpInput[];
     pseudoUrls: PseudoUrlInput[];
+    excludes: GlobInput[];
     linkSelector?: string;
     clickableElementsSelector?: string;
     keepUrlFragments: boolean;
@@ -22,8 +23,10 @@ export interface Input {
     sessionPoolName?: string;
     initialCookies: Parameters<Session['setCookies']>[0];
     useChrome: boolean;
+    maxScrollHeightPixels: number;
     ignoreSslErrors: boolean;
     ignoreCorsAndCsp: boolean;
+    closeCookieModals: boolean;
     downloadMedia: boolean;
     downloadCss: boolean;
     maxRequestRetries: number;