+++ packages/actor-scraper/web-scraper/src/internals/crawler_setup.ts
@@ -22,6 +22,8 @@ import { readFile } from 'node:fs/promises';
 import { HTTPResponse, Page } from 'puppeteer';
 import { dirname } from 'node:path';
 import { fileURLToPath, URL } from 'node:url';
+import { getInjectableScript } from 'idcac-playwright';
+import { setTimeout } from 'node:timers/promises';
 import { createBundle } from './bundle.browser.js';
 import { BreakpointLocation, CHROME_DEBUGGER_PORT, Input, ProxyRotation, RunMode } from './consts.js';
 import { GlobalStore } from './global_store.js';
@@ -408,6 +410,16 @@ export class CrawlerSetup implements CrawlerSetupOptions {
             await page.evaluate(async () => { debugger; }); // eslint-disable-line no-debugger
         }
 
+        if (this.input.closeCookieModals) {
+            await setTimeout(500);
+            await page.evaluate(getInjectableScript());
+            await setTimeout(2000);
+        }
+
+        if (this.input.maxScrollHeightPixels > 0) {
+            await crawlingContext.infiniteScroll({ maxScrollHeight: this.input.maxScrollHeightPixels });
+        }
+
         const startUserFn = process.hrtime();
 
         const namespace = pageContext.apifyNamespace;
@@ -504,6 +516,7 @@ export class CrawlerSetup implements CrawlerSetupOptions {
             selector: this.input.linkSelector,
             globs: this.input.globs,
             pseudoUrls: this.input.pseudoUrls,
+            exclude: this.input.excludes,
             transformRequestFunction: (requestOptions) => {
                 requestOptions.userData ??= {};
                 requestOptions.userData[META_KEY] = {